+++ datafusion/core/src/logical_plan/mod.rs
@@ -21,35 +21,33 @@
 //! Logical query plans can then be optimized and executed directly, or translated into
 //! physical query plans and executed.
 
-pub(crate) mod builder;
 mod expr;
 mod expr_simplier;
 pub mod plan;
 mod registry;
 pub mod window_frames;
-pub use builder::{
-    build_join_schema, union_with_alias, LogicalPlanBuilder, UNNAMED_TABLE,
-};
 pub use datafusion_common::{DFField, DFSchema, DFSchemaRef, ToDFSchema};
 pub use datafusion_expr::{
     expr_fn::binary_expr,
     expr_rewriter,
     expr_visitor::{ExprVisitable, ExpressionVisitor, Recursion},
+    logical_plan::builder::{
+        build_join_schema, union_with_alias, LogicalPlanBuilder, UNNAMED_TABLE,
+    },
     ExprSchemable, Operator,
 };
 pub use expr::{
     abs, acos, and, approx_distinct, approx_percentile_cont, array, ascii, asin, atan,
     avg, bit_length, btrim, call_fn, case, ceil, character_length, chr, coalesce, col,
-    columnize_expr, combine_filters, concat, concat_expr, concat_ws, concat_ws_expr, cos,
-    count, count_distinct, create_udaf, create_udf, date_part, date_trunc, digest,
-    exists, exp, exprlist_to_fields, floor, in_list, in_subquery, initcap, left, length,
-    lit, lit_timestamp_nano, ln, log10, log2, lower, lpad, ltrim, max, md5, min,
-    not_exists, not_in_subquery, now, now_expr, nullif, octet_length, or, power, random,
-    regexp_match, regexp_replace, repeat, replace, reverse, right, round, rpad, rtrim,
-    scalar_subquery, sha224, sha256, sha384, sha512, signum, sin, split_part, sqrt,
-    starts_with, strpos, substr, sum, tan, to_hex, to_timestamp_micros,
-    to_timestamp_millis, to_timestamp_seconds, translate, trim, trunc, unalias, upper,
-    when, Column, Expr, ExprSchema, Literal,
+    combine_filters, concat, concat_expr, concat_ws, concat_ws_expr, cos, count,
+    count_distinct, create_udaf, create_udf, date_part, date_trunc, digest, exists, exp,
+    floor, in_list, in_subquery, initcap, left, length, lit, lit_timestamp_nano, ln,
+    log10, log2, lower, lpad, ltrim, max, md5, min, not_exists, not_in_subquery, now,
+    now_expr, nullif, octet_length, or, power, random, regexp_match, regexp_replace,
+    repeat, replace, reverse, right, round, rpad, rtrim, scalar_subquery, sha224, sha256,
+    sha384, sha512, signum, sin, split_part, sqrt, starts_with, strpos, substr, sum, tan,
+    to_hex, to_timestamp_micros, to_timestamp_millis, to_timestamp_seconds, translate,
+    trim, trunc, unalias, upper, when, Column, Expr, ExprSchema, Literal,
 };
 pub use expr_rewriter::{
     normalize_col, normalize_col_with_schemas, normalize_cols, replace_col,
