+++ datafusion/core/src/optimizer/simplify_expressions.rs
@@ -25,13 +25,13 @@ use crate::logical_plan::{
     LogicalPlan, RewriteRecursion, SimplifyInfo,
 };
 use crate::optimizer::optimizer::OptimizerRule;
-use crate::optimizer::utils;
 use crate::physical_plan::planner::create_physical_expr;
 use crate::scalar::ScalarValue;
 use crate::{error::Result, logical_plan::Operator};
 use arrow::array::new_null_array;
 use arrow::datatypes::{DataType, Field, Schema};
 use arrow::record_batch::RecordBatch;
+use datafusion_expr::utils::from_plan;
 use datafusion_expr::Volatility;
 
 /// Provides simplification information based on schema and properties
@@ -234,7 +234,7 @@ impl OptimizerRule for SimplifyExpressions {
             })
             .collect::<Result<Vec<_>>>()?;
 
-        utils::from_plan(plan, &expr, &new_inputs)
+        from_plan(plan, &expr, &new_inputs)
     }
 }
